# docs/api/openapi.json
{
  "components": {
    "headers": {
      "IdempotentReplay": {
        "description": "Present (value `true`) when this response was replayed from a stored `Idempotency-Key`.",
        "schema": {
          "const": "true",
          "type": "string"
        }
      },
      "RetryAfter": {
        "description": "Seconds to wait before retrying.",
        "schema": {
          "minimum": 0,
          "type": "integer"
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "description": "Any string up to 200 characters, unique per request you intend once (a UUID is fine). Honoured on POST for 24 hours per tenant: a replay with the same key AND the same method, URL and body returns the original response — same status, same body — with `Idempotent-Replay: true`. The same key with a different request is a 422 `VALIDATION_FAILED`. Responses ≥ 500 are not stored, so a failed request may be retried with the same key.",
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "schema": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Malformed JSON, unsupported media type, or a body over 256 KiB (Fastify `FST_ERR_*` code in the envelope)."
      },
...
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`FORBIDDEN` — the key lacks the scope, the request IP is outside the key's allow-list (E-70), a public key was used from an Origin outside its domain list or for anything but `lead_callback`, or the account is suspended/uninstalled."
      },
      "Internal": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`INTERNAL` — unexpected error; quote `request_id` to support. Nothing internal is disclosed."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`NOT_FOUND` — the id does not exist, or belongs to another tenant (row-level security makes those indistinguishable)."
      },
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RateLimitError"
                },
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        },
        "description": "`RATE_LIMITED`. The per-minute limiter answers with the flat `RateLimitError` body; the per-key daily intent cap (E-70) answers with the standard `Error` envelope and `Retry-After: 3600`.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        }
      },
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`UNAUTHENTICATED` — missing, malformed, unknown or revoked API key. Keys look like `nrd_live_…`, `nrd_test_…` or `nrd_pk_…`."
      },
      "ValidationFailed": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`VALIDATION_FAILED` — the body did not match the schema (`details` lists `{path, message}` issues), a phone number could not be parsed, a timestamp is in the future, or an `Idempotency-Key` was reused with a different request."
      }
    },
    "schemas": {
...
      "Error": {
        "description": "The error envelope every non-2xx response uses, except the per-minute rate limiter (see RateLimitError).",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Machine-readable code. Status by code: GATED → 409, ENGINE_UNAVAILABLE → 503, IDEMPOTENT_REPLAY → 200, SIGNATURE_INVALID → 401, VALIDATION_FAILED → 422, UNAUTHENTICATED → 401, FORBIDDEN → 403, RATE_LIMITED → 429, NOT_FOUND → 404, DISPATCH_UNCERTAIN → 500, INTERNAL → 500. Fastify's own client errors (malformed JSON, unsupported media type, body over 256 KiB) use Fastify's `FST_ERR_*` codes with the same envelope.",
                "enum": [
                  "GATED",
                  "ENGINE_UNAVAILABLE",
                  "IDEMPOTENT_REPLAY",
                  "SIGNATURE_INVALID",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "RATE_LIMITED",
                  "NOT_FOUND",
                  "DISPATCH_UNCERTAIN",
                  "INTERNAL"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Optional. For VALIDATION_FAILED from the schema: an array of `{path, message}` issues. Otherwise the error context (flat key/values), when any.",
                "type": "object"
              },
              "message": {
                "description": "Human-readable, safe to show to an operator; never contains a phone number.",
                "type": "string"
              },
              "request_id": {
                "description": "Echo of the request id (`x-cloud-trace-context` when present). Quote it in support requests.",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MerchantWebhook": {
        "properties": {
          "active": {
            "description": "False once deleted, or auto-disabled after 20 consecutive delivery failures.",
            "type": "boolean"
          },
          "consecutive_failures": {
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "disabled_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "events": {
            "items": {
              "enum": [
                "intent.scheduled",
                "intent.gated",
                "intent.cancelled",
                "call.started",
                "call.completed",
                "outcome.final",
                "suppression.created",
                "ticket.created",
                "ticket.resolved",
                "order.cancellation_requested",
                "order.cancelled_by_agent",
                "order.confirmed_by_caller",
                "inbound.call_refused",
                "complaint.received",
                "tenant.paused",
                "erasure.completed",
                "billing.status_changed",
                "billing.capped",
                "billing.approaching_cap"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "url": {
            "description": "https only.",
            "format": "uri",
            "type": "string"
          },
          "webhook_id": {
            "description": "26-character ULID with a type prefix",
            "type": "string"
          }
        },
        "required": [
          "webhook_id",
          "url",
          "events",
          "active",
          "consecutive_failures",
          "disabled_reason",
          "created_at"
        ],
        "type": "object"
      },
...
      "RateLimitError": {
        "description": "The flat body of the per-minute limiter (@fastify/rate-limit). The per-key DAILY intent cap answers 429 with the standard Error envelope instead.",
        "properties": {
          "code": {
            "const": "RATE_LIMITED",
            "type": "string"
          },
          "error": {
            "const": "Too Many Requests",
            "type": "string"
          },
          "message": {
            "description": "e.g. \"rate limit exceeded, retry in 12s\"",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "statusCode": {
            "const": 429,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "publicSiteKey": {
        "bearerFormat": "nrd_pk_…",
        "description": "Public site key for browsers: `POST /v1/intents` with `use_case: lead_callback` only, from an allow-listed Origin, 10 requests/minute per IP.",
        "scheme": "bearer",
        "type": "http"
      },
      "secretKey": {
        "bearerFormat": "nrd_live_… / nrd_test_…",
        "description": "Secret API key. The scope names listed on each operation are what the key must carry.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Naaradh",
      "url": "https://naaradh.com"
    },
...
    "summary": "Two-way AI voice agent for commerce — outbound call intents and the inbound support line.",
    "title": "Naaradh API",
    "version": "0.0.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/webhooks": {
      "get": {
        "description": "Every endpoint registered for your account, active or not, newest first. Secrets are never included.\n\nScope: `webhooks:read`.",
        "operationId": "listWebhooks",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "webhooks": {
                      "items": {
                        "$ref": "#/components/schemas/MerchantWebhook"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "webhooks"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Endpoints."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "webhooks:read"
            ]
          }
        ],
        "summary": "List webhook endpoints",
        "tags": [
          "Merchant webhooks"
        ]
      },
      "post": {
        "description": "Registers an HTTPS endpoint for a set of event types. The signing `secret` (`whsec_…`) is returned exactly once, in this response — Naaradh stores only a reference to it. Every delivery is signed with it (see the `webhooks` section: `X-Naaradh-Signature: t=<unix>,v1=<hex hmac_sha256(secret, t + \".\" + body)>`, 5-minute replay window).\n\nScope: `webhooks:write`.",
        "operationId": "createWebhook",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "events": {
                    "description": "Event types to deliver to this endpoint.",
                    "items": {
                      "enum": [
                        "intent.scheduled",
                        "intent.gated",
                        "intent.cancelled",
                        "call.started",
                        "call.completed",
                        "outcome.final",
                        "suppression.created",
                        "ticket.created",
                        "ticket.resolved",
                        "order.cancellation_requested",
                        "order.cancelled_by_agent",
                        "order.confirmed_by_caller",
                        "inbound.call_refused",
                        "complaint.received",
                        "tenant.paused",
                        "erasure.completed",
                        "billing.status_changed",
                        "billing.capped",
                        "billing.approaching_cap"
                      ],
                      "type": "string"
                    },
                    "minItems": 1,
                    "type": "array"
                  },
                  "url": {
                    "description": "Must be `https://`. Redirects are not followed.",
                    "maxLength": 2048,
                    "type": "string"
                  }
                },
                "required": [
                  "url",
                  "events"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "events": {
                      "items": {
                        "enum": [
                          "intent.scheduled",
                          "intent.gated",
                          "intent.cancelled",
                          "call.started",
                          "call.completed",
                          "outcome.final",
                          "suppression.created",
                          "ticket.created",
                          "ticket.resolved",
                          "order.cancellation_requested",
                          "order.cancelled_by_agent",
                          "order.confirmed_by_caller",
                          "inbound.call_refused",
                          "complaint.received",
                          "tenant.paused",
                          "erasure.completed",
                          "billing.status_changed",
                          "billing.capped",
                          "billing.approaching_cap"
                        ],
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "note": {
                      "const": "store the secret now; it is not retrievable later",
                      "type": "string"
                    },
                    "secret": {
                      "description": "`whsec_` + 32 url-safe base64 characters. Not retrievable later; delete and re-create to rotate.",
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "webhook_id": {
                      "description": "26-character ULID with a type prefix",
                      "type": "string"
                    }
                  },
                  "required": [
                    "webhook_id",
                    "url",
                    "events",
                    "secret",
                    "note"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Registered. Store `secret` now.",
            "headers": {
              "Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "webhooks:write"
            ]
          }
        ],
        "summary": "Register a webhook endpoint",
        "tags": [
          "Merchant webhooks"
        ]
      }
    },
...
    "/v1/webhooks/{id}": {
      "delete": {
        "description": "Deactivates the endpoint (kept for the delivery history, `disabled_reason: \"deleted via api\"`). Pending deliveries to it are dead-lettered. 404 when it does not exist or is already inactive.\n\nScope: `webhooks:write`.",
        "operationId": "deleteWebhook",
        "parameters": [
          {
            "description": "Webhook id (`mwh_…`).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "webhooks:write"
            ]
          }
        ],
        "summary": "Delete a webhook endpoint",
        "tags": [
          "Merchant webhooks"
        ]
      }
    }
  },
  "security": [
    {
      "secretKey": []
    }
  ],
  "servers": [
    {
      "description": "Production",
      "url": "https://api.naaradh.com"
    },
    {
      "description": "Local development (apps/api default PORT)",
      "url": "http://localhost:3001"
    }
  ],
...
}